requirements.txt
accelerate==0.22.0
bitsandbytes==0.41.1
datasets==2.14.5
evaluate==0.4.0
matplotlib==3.7.2
nltk==3.8.1
numpy==1.24.4
pandas==2.2.2
peft==0.5.0
py-readability-metrics==1.4.5
sacrebleu==2.3.1
sacremoses==0.0.53
scipy==1.11.2
sentencepiece==0.1.99
spacy==3.6.1
torch==2.0.1
tqdm==4.66.4
transformers==4.33.1
typed_argument_parser==1.8.1
unbabel-comet==2.0.2
